+++ MNO/proyecto_final/MNO_2017/proyectos/equipos/equipo_5/code/matrixVectorMult.cu
@@ -2,10 +2,10 @@
 #include <iostream>
 #include <vector>
 
-__global__ void gaxpy(int *y, int *a, int *x, int m, int n){
+__global__ void gaxpy(double *y, double *a, double *x, int m, int n){
 	int bid = blockIdx.x;
         int tid = threadIdx.x;
-	extern __shared__ int dots_s[];
+	extern __shared__ double dots_s[];
 	if(bid<m)
 		if(tid<n){
 
@@ -22,31 +22,31 @@ __global__ void gaxpy(int *y, int *a, int *x, int m, int n){
 	}
 }
 
-std::vector<int> matrixVectorMultiplication(int* a, int mc, int nc, int* x){
+std::vector<double> matrixVectorMultiplication(double* a, int mc, int nc, double* x){
 	int* m, *n;
 	m = &mc;
 	n = &nc;
-	std::vector<int> y(mc,0);
-	int *device_y, *device_a, *device_x;
+	std::vector<double> y(mc,0);
+	double *device_y, *device_a, *device_x;
 	int *device_m, *device_n;
 	//alojando en device
-	cudaMalloc((void **)&device_y, sizeof(int)*mc);
-	cudaMalloc((void **)&device_a, sizeof(int)*mc*nc);
-	cudaMalloc((void **)&device_x, sizeof(int)*nc);
+	cudaMalloc((void **)&device_y, sizeof(double)*mc);
+	cudaMalloc((void **)&device_a, sizeof(double)*mc*nc);
+	cudaMalloc((void **)&device_x, sizeof(double)*nc);
 	cudaMalloc((void **)&device_m, sizeof(int));
 	cudaMalloc((void **)&device_n, sizeof(int));
 	//copiamos arreglos a, x a la GPU
-	cudaMemcpy(device_a,a,mc*nc*sizeof(int), cudaMemcpyHostToDevice);
-	cudaMemcpy(device_x,x,nc*sizeof(int), cudaMemcpyHostToDevice);
-	cudaMemcpy(device_y,y.data(),mc*sizeof(int), cudaMemcpyHostToDevice);
+	cudaMemcpy(device_a,a,mc*nc*sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(device_x,x,nc*sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(device_y,y.data(),mc*sizeof(double), cudaMemcpyHostToDevice);
 	cudaMemcpy(device_m,m,sizeof(int), cudaMemcpyHostToDevice);
 	cudaMemcpy(device_n,n,sizeof(int), cudaMemcpyHostToDevice);
 	//mandamos a llamar a suma_vect:
-	gaxpy<<<mc,nc,sizeof(int)*mc*nc>>>(device_y,device_a,device_x,mc,nc);
+	gaxpy<<<mc,nc,sizeof(double)*mc*nc>>>(device_y,device_a,device_x,mc,nc);
 //	for(unsigned i=0; i<y.size();i++)
 //		std::cout << "yi[i] = " << y[i] << "\n";
 	//copia del resultado al arreglo y:
-	cudaMemcpy(y.data(),device_y,mc*sizeof(int),cudaMemcpyDeviceToHost);
+	cudaMemcpy(y.data(),device_y,mc*sizeof(double),cudaMemcpyDeviceToHost);
 //	for(unsigned i=0; i<y.size();i++)
 //		std::cout << "yf[i] = " << y[i] << "\n";
 	cudaFree(device_y);
